+++ packages/x6/src/connection/marker/cross.ts
@@ -1,4 +1,5 @@
 import { Path } from '../../geometry'
+import { normalizeMarker } from './util'
 import { Marker } from './index'
 
 export interface CrossMarkerOptions {
@@ -19,6 +20,6 @@ export const cross: Marker.Definition<CrossMarkerOptions> = (options) => {
   return {
     type: 'path',
     fill: 'none',
-    d: Marker.normalize(path.serialize(), options.offset || -width / 2),
+    d: normalizeMarker(path.serialize(), options.offset || -width / 2),
   }
 }
