+++ packages/x6/src/view/cell.ts
@@ -1,8 +1,8 @@
+import { Registry } from '../common'
 import { Nilable, KeyValue } from '../types'
 import { Rectangle, Point } from '../geometry'
 import { ArrayExt, ObjectExt, Dom } from '../util'
-import { Registry } from '../common'
-import { Attr } from '../definition'
+import { Attr } from '../definition/attr'
 import { Cell } from '../model/cell'
 import { Edge } from '../model/edge'
 import { Model } from '../model/model'
