+++ packages/x6/src/view/tool.ts
@@ -1,6 +1,6 @@
 import { KeyValue } from '../types'
 import { Dom, ObjectExt, StringExt } from '../util'
-import { NodeTool, EdgeTool } from '../definition'
+import { NodeTool, EdgeTool } from '../definition/tool'
 import { View } from './view'
 import { CellView } from './cell'
 import { Markup } from './markup'
